refactor: move playmode button out
