+++ src/components/player/PlayPauseButton.tsx
@@ -23,7 +23,7 @@ export const PlayPauseButton: React.FC<{
   const showPause = playWhenReady && !(isErrored || isEnded);
   const showBuffering = playWhenReady && isLoading;
   return showBuffering ? (
-    <ActivityIndicator size={50} />
+    <ActivityIndicator size={55} />
   ) : (
     <IconButton
       icon={showPause ? 'pause' : 'play'}
